Rename and split Array/Iterator APIs
