+++ numpy-sys/src/lib.rs
@@ -4,6 +4,5 @@ extern crate python3_sys as pyffi;
 
 pub mod types;
 pub mod array;
-pub mod arrayiter;
+pub mod iterator;
 pub mod ufunc;
-pub mod multiarray;
